requirements.txt
loguru>=0.7.0
playwright>=1.40.0
playwright-stealth>=1.0.0
pytest>=8.0.0
google-generativeai>=0.8.0
google-genai>=1.0.0
APScheduler>=3.10.0
streamlit>=1.32.0
requests>=2.31.0
